Refactored tasks in correspondence with introduction of portfolio
